use int for ms values

+++ src/DotPulsar/ConsumerOptions.cs
@@ -106,11 +106,11 @@ namespace DotPulsar
         /// Delay to wait before redelivering messages that failed to be processed.
         /// When an application uses IConsumer.NegativeAcknowledge(Message), failed messages are redelivered after a fixed timeout.
         /// </summary>
-        public ulong NegativeAckRedeliveryDelayMicros { get; set; }
+        public int NegativeAckRedeliveryDelayMicros { get; set; }
 
         /// <summary>
         /// Timeout of unacked messages
         /// </summary>
-        public ulong AckTimeoutMillis { get; set; }
+        public int AckTimeoutMillis { get; set; }
     }
 }
